delete confirmation, ingredients and instructions added

+++ src/_root/pages/PostDetails.tsx
@@ -28,9 +28,13 @@ const PostDetails = () => {
   );
 
   const handleDeletePost = () => {
-    deletePost({ postId: id, imageId: post?.imageId });
-    navigate(-1);
+    const isConfirmed = window.confirm("Are you sure you want to delete this post?");
+    if (isConfirmed) {
+      deletePost({ postId: id, imageId: post?.imageId });
+      navigate(-1);
+    }
   };
+  
 
   return (
     <div className="post_details-container">
@@ -130,6 +134,32 @@ const PostDetails = () => {
                 ))}
               </ul>
             </div>
+             
+            <div className="flex flex-col flex-1 w-full small-medium lg:base-regular">
+  <p>Ingredients</p>
+  <ul className="flex flex-col gap-1 mt-2">
+    {post?.ingredients.map((ingredient: string, index: number) => (
+      <li key={index} className="">
+        {/* Add bullet point before each ingredient */}
+        &#8226; {ingredient}
+      </li>
+    ))}
+  </ul>
+</div>
+
+
+            <div className="flex flex-col flex-1 w-full small-medium lg:base-regular">
+              <p>Instructions</p>
+              <ul className="flex flex-col gap-1 mt-2">
+                {post?.instructions.map((instruction: string, index: number) => (
+                  <li
+                    key={`${instruction}${index}`}
+                    className="">
+                    {`${index + 1}. ${instruction}`}
+                  </li>
+                ))}
+              </ul>
+            </div>
 
             <div className="w-full">
               <PostStats post={post} userId={user.id} />
@@ -142,7 +172,7 @@ const PostDetails = () => {
         <hr className="border w-full border-dark-4/80" />
 
         <h3 className="body-bold md:h3-bold w-full my-10">
-          More Related Posts
+          More From This Cook
         </h3>
         {isUserPostLoading || !relatedPosts ? (
           <Loader />
